requirements.txt
ruff
python-dotenv>=1.0.0
pre-commit==3.8.0
pytest==8.3.3
pytest-asyncio==0.23.6
pyautogui==0.9.54
pyperclip>=1.8.0
streamlit>=1.38.0
jsonschema==4.22.0
boto3>=1.28.57
google-auth<3,>=2
gradio>=5.18.0
screeninfo
uiautomation

# make sure to install the correct version of torch (cuda, mps, cpu, etc.)
# torch
# torchvision

transformers
accelerate
dashscope
huggingface_hub
openai

# 飞书网关 (可选)
lark-oapi>=1.2.0
